backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
pysmashgg==1.2.0.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==1.26.16
